move fetch and create releases to sharedResources
